refactor: extract shared FACTION_COLORS, formatCompactDuration, eventKey
