+++ src/features/archives/eventKey.mjs
@@ -1,0 +1,3 @@
+export function eventKey(event) {
+    return `${event.type}-${event.event_id}`;
+}
